+++ src/tepollwebsocket.cpp
@@ -71,7 +71,7 @@ bool TEpollWebSocket::isBinaryRequest() const
 
 void TEpollWebSocket::sendTextForPublish(const QString &text, const QObject *except)
 {
-    tSystemDebug("sendText  text len:%lld  (pid:%d)", text.length(), (int)QCoreApplication::applicationPid());
+    tSystemDebug("sendText  text len:%lld  (pid:%d)", (qint64)text.length(), (int)QCoreApplication::applicationPid());
     if (except != this) {
         TAbstractWebSocket::sendText(text);
     }
@@ -80,7 +80,7 @@ void TEpollWebSocket::sendTextForPublish(const QString &text, const QObject *exc
 
 void TEpollWebSocket::sendBinaryForPublish(const QByteArray &binary, const QObject *except)
 {
-    tSystemDebug("sendBinary  binary len:%lld  (pid:%d)", binary.length(), (int)QCoreApplication::applicationPid());
+    tSystemDebug("sendBinary  binary len:%lld  (pid:%d)", (qint64)binary.length(), (int)QCoreApplication::applicationPid());
     if (except != this) {
         TAbstractWebSocket::sendBinary(binary);
     }
@@ -89,7 +89,7 @@ void TEpollWebSocket::sendBinaryForPublish(const QByteArray &binary, const QObje
 
 void TEpollWebSocket::sendPong(const QByteArray &data)
 {
-    tSystemDebug("sendPong  data len:%lld  (pid:%d)", data.length(), (int)QCoreApplication::applicationPid());
+    tSystemDebug("sendPong  data len:%lld  (pid:%d)", (qint64)data.length(), (int)QCoreApplication::applicationPid());
     TAbstractWebSocket::sendPong(data);
 }
