+++ src/tredisdriver.cpp
@@ -41,7 +41,7 @@ bool TRedisDriver::request(const QByteArrayList &command, QVariantList &response
 
     for (;;) {
         if (!readReply()) {
-            tSystemError("Redis read error   pos:%d  buflen:%lld", _pos, _buffer.length());
+            tSystemError("Redis read error   pos:%d  buflen:%lld", _pos, (qint64)_buffer.length());
             close();
             break;
         }
